move Bandpass to catalog tools; add LSST filters
